Improved resolution for RGB/CW color and brightness [Dev note: Will merge in few days, just need final feedback] (#447)
@rogere66 better LED colours pull request applied; so far in unmodified form, but I will adjust channels handling in the several hours and then do release

+++ src/hal/bk7231/hal_pins_bk7231.c
@@ -118,7 +118,7 @@ void HAL_PIN_PWM_Start(int index) {
 #endif
 	bk_pwm_start(pwmIndex);
 }
-void HAL_PIN_PWM_Update(int index, int value) {
+void HAL_PIN_PWM_Update(int index, float value) {
 	int pwmIndex;
 
 	pwmIndex = PIN_GetPWMIndexForPinIndex(index);
